requirements.txt
maxapi==0.9.7
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
jinja2==3.1.4
aiofiles==24.1.0
loguru==0.7.2
